requirements.txt
notifypy
chime
Pillow
customtkinter
tkextrafont